+++ src/main/java/org/example/Main.java
@@ -10,7 +10,7 @@ import java.util.Map;
  * In der Main Klasse wird die GUI für die To-do-Liste erstellt.
  * */
 public class Main {
-    protected JFrame mainFrame;
+    private JFrame mainFrame;
     private JPanel panel;
 
     /**
@@ -18,37 +18,51 @@ public class Main {
      * Erstellt das Hauptfenster, die Menüleiste und fügt die To-do-Listen ins Hauptfenster.
      * */
     private void createGUI() {
-        mainFrame = new JFrame("To Do");
-        mainFrame.setSize(500,500);
         mainFrame.setLocationRelativeTo(null);
         mainFrame.setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
         mainFrame.setLayout(new GridLayout(0,1));
 
         JPanel controlPanel = new JPanel(new FlowLayout());
-        mainFrame.add(controlPanel);
 
+        // Erstellt die Scrollleisten und fügt diese zum Frame
+        JScrollPane scrollPane = new JScrollPane(
+                controlPanel, JScrollPane.VERTICAL_SCROLLBAR_ALWAYS, JScrollPane.HORIZONTAL_SCROLLBAR_ALWAYS
+        );
+
+        mainFrame.add(scrollPane);
         panel = new JPanel();
         // Erstellung, gestaltung und Hinzufügung des Layouts zum Panel
-        GridLayout gridLayout = new GridLayout(0,5);
+        GridLayout gridLayout = new GridLayout();
         gridLayout.setHgap(20);
         gridLayout.setVgap(20);
         panel.setLayout(gridLayout);
 
         createJMenuBar();
-
-        Map<String, Map<String, Boolean>> data = new DataHandler().getData();
+        Map<String, Map<String, List<Object>>> data = new DataHandler().getData();
         List<String> keysList = new ArrayList<>(data.keySet()); // Liste der Namen der To-do-Listen
 
         // Erstellt die To-do-Listen mit den Elementen
         for (String name : keysList) {
-            new CreateTaskLabel(panel).createJLabel(name);
+            new CreateTaskLabel(panel, mainFrame).createJLabel(name);
         }
 
         controlPanel.add(panel);
 
+        //Passt die columns der Anzahl der Listen an
+        int panelCount = panel.getComponentCount();
+
+        if (panelCount < 5) {
+            gridLayout.setColumns(panelCount);
+        } else {
+            gridLayout.setColumns(5);
+            gridLayout.setRows(0);
+        }
+
         // Wenn es keine Elemente gibt, dann soll das Fenster nicht angepasst werden
-        if (panel.getComponentCount() > 0) {
+        if (panelCount > 0) {
             mainFrame.pack();
+        } else {
+            mainFrame.setSize(300,300);
         }
 
         mainFrame.setVisible(true);
@@ -64,7 +78,7 @@ public class Main {
         JMenuItem close = new JMenuItem("Schließen");
 
         // ActionListener für die Menüitems
-        newList.addActionListener(_ -> new NewTodo(panel));
+        newList.addActionListener(_ -> new NewTodo(panel, mainFrame));
         close.addActionListener(_ -> System.exit(0));
 
         // Fügt die Menüitems dem Dateiitem
@@ -78,6 +92,8 @@ public class Main {
 
     public static void main(String[] args) {
         Main main = new Main();
+        main.mainFrame = new JFrame("To Do");
+        //main.mainFrame.setSize(500,600);
 
         main.createGUI();
     }